Add support for select line on MacOS.
The default for select a word seemed to have been
SUPER, but on my Mac, select a word is ALT, with
select a line being SUPER. This change makes floem
have ALT as select a word, with SUPER as select
a line. Please let me know if this
is not the expected behavior.