Smarter last vline w/ new layout cache
Last vline could grow very expensive (10ms) on large (200k lines) files, because it was forced to look up every (font_size, line) in the hashmap.
(Possibly it was more expensive than 10ms previously, because it had a second-level hashmap rather than an array!)
Now it is quite cheap as the new layout cache makes it simple to only do that iteration on the lines actually in view.

It does have the issue that it isn't respecting different font sizes for each line but that's fixable.