Remove font size cache
This was originally only thought to be used for code lens in Lapce, which would shrink the font size of lines, but I don't believe that's implemented.
As well, it should be fast enough to just rerender those text layouts. This makes many operations much faster, no longer having to deal with a hashmap or the trait object indirection every loop iter.