Expose the scale value from Renderer.
Rationale: It is possible to set the scale on a PaintCx, but not to restore it to its previous value,
and `PaintCx.save()` and `PaintCx.restore()` do not handle it, since it is writing directly to a property
of the underlying renderer.

Use-case: In order to debug layout issues, it is useful to be able to scale a component's content down,
render it that way and observe if there is any painting outside the expected bounds.