deps: kurbo is used via peniko.
Since all usages of `kurbo` are via `peniko`, the direct dependency
on `kurbo` is not needed. This makes it easier to keep versions
in sync.

The re-export of `kurbo` was happening directly so it could have
been out of sync with an updated `peniko`, so re-export the version
from `peniko` instead.

`kurbo` was being used to enable the `serde` feature in it, so do
that now via `peniko` as allowed in v0.1.1.