simplify widget gallery by using list (#683)
As part of this, individual tab selectors are not marked as keyboard navigable but the list is. This makes it so that the items must be navigated to using the list selection when the list has focus. the previous behavior was strange as items could be navigated to separately using tab and the list selection. this led to confusion.