virtual_stack: handle selection based on index (#869)
The ViewId can be lost if the selected item is dropped from the
underlying virtual vector, for example because of scrolling to a
different part of the list. When the item gets recreated (when scrolling
back to it), the ViewId will be different and the item will no longer be
selected.

Conveniently, this allows dropping the `pending_idx` handling.

Fixes: #867

Signed-off-by: Conrad Hoffmann <ch@bitfehler.net>